WiP build attendee registration page Move form into a page.js
